requirements.txt
avro-python3==1.10.2
